1) Добавлена возможность использовать С для любого класса, реализующего интерфейс TestLogging 2) Добавлено предварительное сохранение методов в HashSet в DemoInvocationHandler
